[ENHANCEMENT] Change DashboardCard bg color + Merge CreateDashboardDialogs in one component
