+++ ui/app/src/components/AddProjectDialog/AddProjectDialog.tsx
@@ -26,8 +26,8 @@ export interface AddProjectDialogProps {
 /**
  * Dialog used to create a project.
  * @param props.open Define if the dialog should be opened or not.
- * @param props.closeDialog Provides the function to close itself.
- * @param props.onConfirm Action to perform when user confirmed.
+ * @param props.onClose Provides the function to close itself.
+ * @param props.onSuccess Action to perform when user confirmed.
  * @constructor
  */
 export const AddProjectDialog = (props: AddProjectDialogProps) => {
